4 tab bar icons and journal text view border
Added the 4 tab bar icons in all 3 resolutions. Made a credit document linked at the bottom of the resources view controller. Added border for the journal text view with the same style as that of the journal text field. Added a placeholder text to help get the user in a flow and understand what the text view is for.